+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Add support for SRDF descriptions (thanks to @nickswalker)
 - Description: Add SRDFs for Franka Emika robots: Panda, FER, FR3, FR3v2, FR3v2_1 (thanks to @nickswalker)
 - Description: Add Toyota HSR-B and HSR-C models
+- Description: OpenArm v2 (MJCF) (thanks to @kou)
 - Description: RBY1 Mobile Manipulatior (MJCF)
 - Description: Stretch 4 (URDF) (thanks to @nickswalker)
 - Description: TIAGo official (URDF) (thanks to @nickswalker)